+++ package.json
@@ -26,6 +26,7 @@
     "@appscode/json2yaml": "^0.1.2",
     "@datadog/browser-logs": "^4.25.0",
     "@kong/kongponents": "^8.12.0",
+    "brandi": "^4.0.2",
     "prismjs": "^1.29.0",
     "semver": "^7.3.8",
     "tailwindcss": "^3.2.4",
