+++ src/services/env.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from '@jest/globals'
 
-import { semver } from '@/services/env'
+import { semver } from '@/services/container'
 
 describe('env', () => {
   describe('semver', () => {
